requirements.txt
requests~=2.33.1
python-dotenv~=1.2.2
aiogram~=3.27.0
playwright~=1.58.0